requirements.txt
Flask
requests
Pillow
werkzeug